Added code patch to the code base - Version number bumped to 2.04
The patch is a fix to the custom code base.  The fix adjusts the device
so that when the device receives a message of the form ..XPOLLS<last 6
of imei here> it puts itself in a state where if it receives another
XPOLLS message before it responds to the first that it won't go into a
bugged state where it starts reporting every 5 seconds.
Version number is now 2.04.